+++ test/integration/failure/failure.spec.js
@@ -2,9 +2,9 @@ const path = require('path');
 const chai = require('chai');
 const expect = chai.expect;
 
-const parser = require('../../index');
+const parser = require('../../../index');
 
-describe('Failure', () => {
+describe('SvelteDoc - Failure', () => {
     it('Component have invalid JS syntax should fall into catch block', (done) => {
         parser.parse({
             filename: path.resolve(__dirname, 'failure.invalid.js.svelte'),