+++ test/integration/overall/overall.main.spec.js
@@ -4,9 +4,9 @@ const expect = chai.expect;
 
 const fs = require('fs');
 
-const parser = require('../../index');
+const parser = require('../../../index');
 
-describe('Overall', () => {
+describe('SvelteDoc - Overall', () => {
     it('Component should be parsed with all features', (done) => {
         parser.parse({
             filename: path.resolve(__dirname, 'main.svelte'),